pyproject.toml
[project]
name = "etl-pipeline-airflow"
version = "0.1.0"
description = "E-Commerce End-to-End ELT Pipeline with Airflow, dbt, and PostgreSQL"
authors = [
    { name = "Malik Fajar Mubarok", email = "mubarok.malikfajar@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.5",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "great_expectations>=1.10.0",
    "dbt-core>=1.10.0",
    "dbt-postgres>=1.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "faker>=22.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "ty>=0.0.7",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "ty>=0.0.7",
]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "ARG", "PLE", "PLR", "PLW", "RUF"]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "W293",  # Blank line contains whitespace (docstring)
    "PT011",  # Some tests need to use broad exception
    "B017",  # Some tests need to use broad exception
    "DTZ007",  # Date string to datetime (without timezone)
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "scripts/seed_source_db.py",
    "scripts/run_extract_locally.py",
    "scripts/run_load_locally.py",
    "scripts/run_validate_locally.py",
]

line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["scripts", "dags"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=scripts",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.ty.environment]
python = ".venv"
python-version = "3.12"
root = ["./"]

[tool.ty.src]
exclude = ["./dags", "./dbt_project", "./tests"]

[tool.ty.rules]
deprecated = "warn"
index-out-of-bounds = "error"
invalid-argument-type = "error"
invalid-assignment = "error"
invalid-attribute-access = "error"
invalid-declaration = "error"
invalid-key = "error"
invalid-return-type = "error"
missing-argument = "error"
no-matching-overload = "error"
non-subscriptable = "error"
not-iterable = "error"
possibly-missing-attribute = "warn"
possibly-missing-implicit-call = "warn"
possibly-unresolved-reference = "warn"
too-many-positional-arguments = "error"
unknown-argument = "error"
unresolved-attribute = "error"
unresolved-import = "error"
unresolved-reference = "error"
unsupported-operator = "error"

[tool.coverage.run]
source = ["scripts"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
